requirements.txt
# Core
streamlit>=1.42.0
streamlit-flow-component>=1.6.1
python-dotenv>=1.0.0

# LangChain
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
openai>=1.30.0

# Data processing
pandas>=2.2.0
polars>=0.20.0
pyarrow>=18.0.0
duckdb>=1.1.0
openpyxl>=3.1.0
numpy>=1.26.0

# SQL
sqlglot>=23.0.0

# Google Cloud
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.10.0
google-auth>=2.35.0
db-dtypes>=1.3.0
requests>=2.31.0

# Charts
plotly>=5.20.0

# Audit log
jsonlines>=4.0.0

# Utilities
zipfile36>=0.1.3
